REF: make a dedicated pymc helper module
